uh liked names should theoretically work?

+++ android_stuff/app/src/main/java/com/example/sasha/finalsoftware/ExampleAdapter.java
@@ -74,7 +74,7 @@ public class ExampleAdapter extends RecyclerView.Adapter<ExampleAdapter.ExampleV
         holder.tvName.setText(currentItem.getmName());
         holder.tvSex.setText(currentItem.getmSex());
         holder.tvYear.setText(currentItem.getmYear());
-        holder.tvPercent.setText(currentItem.getmPercent());
+        holder.tvPercent.setText((int) currentItem.getmPercent());
 
     }
